backend/requirements.txt
# CineMate Care - Python Dependencies
# Install with: pip install -r requirements.txt

# Azure AI Services
azure-ai-vision-imageanalysis>=1.0.0    # Azure AI Vision for scene analysis
openai>=1.0.0                            # Azure OpenAI SDK for GPT-4o
azure-cognitiveservices-speech>=1.34.0   # Azure Speech TTS/STT

# Computer Vision & Image Processing
opencv-python>=4.8.0                     # Webcam capture and frame processing
scikit-image>=0.22.0                     # SSIM for keyframe detection
numpy>=1.24.0                            # Array operations
Pillow>=10.0.0                           # Image handling

# Web Dashboard
streamlit>=1.29.0                        # Real-time dashboard UI
plotly>=5.18.0                           # Interactive charts

# Utilities
python-dotenv>=1.0.0                     # Environment variable management
tenacity>=8.2.0                          # Retry logic for API calls
aiohttp>=3.9.0                           # Async HTTP for API calls

# Competition Demo Features (Phase 6-8)
pyaudio>=0.2.14                          # Audio level monitoring
keyboard>=0.13.5                         # Wizard of Oz hotkey backup
